requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.11.3
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.12.1
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
cachelib==0.9.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
cssselect2==0.7.0
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.7
dnspython==2.4.2
elastic-transport==8.4.0
elasticsearch==8.10.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
filelock==3.12.2
Flask==2.3.3
flask-babel==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-OAuthlib==0.9.6
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.42.1
frozenlist==1.4.0
gunicorn==20.1.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httpie==3.2.2
idna==3.4
imageio==2.31.3
implicit==0.7.2
importlib-resources==6.1.0
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
limits==3.6.0
livereload==2.6.3
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python==8.1.0
nltk==3.8.1
numpy==1.26.4
oauthlib==2.1.0
opencv-python-headless==4.8.0.76
ordered-set==4.1.0
packaging==23.1
pandas==2.2.1
Pillow==9.5.0
platformdirs==3.10.0
protobuf==4.21.12
psycopg2==2.9.9
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyngrok==7.1.3
pyphen==0.14.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.8.0
python-multipart==0.0.6
python-slugify==8.0.1
python-socketio==5.10.0
python3-memcached==1.51
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.1.0
requests-toolbelt==1.0.0
rich==13.5.3
rpds-py==0.10.6
schedule==1.2.1
scipy==1.12.0
simple-websocket==1.0.0
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
SQLAlchemy==2.0.20
SQLAlchemy-Utils==0.41.1
text-unidecode==1.3
threadpoolctl==3.3.0
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
twilio==8.10.0
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.16
validators==0.21.2
virtualenv==20.24.3
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
yarl==1.9.2
zopfli==0.2.2
